+++ services/audiopolicy/service/Spatializer.cpp
@@ -880,7 +880,7 @@ void Spatializer::onActualModeChangeMsg(HeadTrackingMode mode) {
                     spatializerMode = HeadTracking::Mode::RELATIVE_SCREEN;
                     break;
                 default:
-                    LOG_ALWAYS_FATAL("Unknown mode: %d", mode);
+                    LOG_ALWAYS_FATAL("Unknown mode: %d", static_cast<int>(mode));
             }
         }
         mActualHeadTrackingMode = spatializerMode;
